Added Brazilian Portuguese Language
Translation of file to Brazilian Portuguese.
Note: I did not compile and test it please, do review if i did not break anything. Another point to make is, Brazilian Portuguese is usually pointed as Pt-br not Pt, this is for European Portuguese (from Portugal). But I'm not sure where to change it properly.

Please review.